fix(layout): make shared updates conflict-safe
