+++ apps/web/src/shell/layoutSync/index.ts
@@ -1,0 +1,1 @@
+export * from "./layoutSync";
