+++ apps/web/src/shell/layoutSync/layoutSync.test.ts
@@ -1,17 +1,17 @@
 import { beforeEach, describe, expect, test } from "bun:test";
 import type {
-	LayoutChangedPayload,
 	LayoutReplaceParams,
+	LayoutReplaceResult,
 	WorkspaceLayoutDocument,
 	WorkspaceLayoutSnapshot,
 } from "@thinkrail/contracts";
+import { useAppStore } from "../../store";
 import {
 	commitWorkspaceLayout,
 	hydrateWorkspaceLayout,
 	resetLayoutSyncForTests,
 	setLayoutReplaceRequesterForTests,
-} from "../shell/layoutSync";
-import { useAppStore } from "./appStore";
+} from "./index";
 
 function document(tabId: string): WorkspaceLayoutDocument {
 	return {
@@ -30,6 +30,12 @@ function document(tabId: string): WorkspaceLayoutDocument {
 function snapshot(revision: number, tabId: string): WorkspaceLayoutSnapshot {
 	return { workspaceId: "ws", revision, document: document(tabId) };
 }
+
+type PendingRequest = {
+	params: LayoutReplaceParams;
+	resolve: (result: LayoutReplaceResult) => void;
+	reject: (error: Error) => void;
+};
 
 beforeEach(() => {
 	resetLayoutSyncForTests();
@@ -51,14 +57,18 @@ describe("synchronized layout store", () => {
 		store.installLayoutSnapshot(snapshot(1, "accepted"));
 		store.beginLayoutCommit("ws", document("first"), "m1");
 		store.beginLayoutCommit("ws", document("second"), "m2");
-		expect(useAppStore.getState().layoutDocumentsByWorkspace.ws).toEqual(document("second"));
+		const optimistic = useAppStore.getState();
+		expect(optimistic.layoutDocumentsByWorkspace.ws).toEqual(document("second"));
+		expect(optimistic.layoutPendingByWorkspace.ws?.map((write) => write.expectedRevision)).toEqual([
+			1, 2,
+		]);
 
 		store.installLayoutSnapshot(snapshot(2, "first"), "m1");
 		const afterFirst = useAppStore.getState();
 		expect(afterFirst.layoutSnapshotsByWorkspace.ws).toEqual(snapshot(2, "first"));
 		expect(afterFirst.layoutDocumentsByWorkspace.ws).toEqual(document("second"));
-		expect(afterFirst.layoutPendingByWorkspace.ws?.map((write) => write.mutationId)).toEqual([
-			"m2",
+		expect(afterFirst.layoutPendingByWorkspace.ws).toMatchObject([
+			{ mutationId: "m2", expectedRevision: 2 },
 		]);
 
 		store.installLayoutSnapshot(snapshot(3, "second"), "m2");
@@ -105,11 +115,6 @@ describe("synchronized layout store", () => {
 	});
 
 	test("serializes writes per workspace and prevents a rolled-back dependent write reaching the host", async () => {
-		type PendingRequest = {
-			params: LayoutReplaceParams;
-			resolve: (payload: LayoutChangedPayload) => void;
-			reject: (error: Error) => void;
-		};
 		const requests: PendingRequest[] = [];
 		setLayoutReplaceRequesterForTests(
 			(params) =>
@@ -136,16 +141,20 @@ describe("synchronized layout store", () => {
 		const independent = commitWorkspaceLayout("other", document("other-next"));
 		await Bun.sleep(0);
 		expect(requests.map((request) => request.params.workspaceId)).toEqual(["ws", "other"]);
+		expect(requests.map((request) => request.params.expectedRevision)).toEqual([1, 1]);
 
 		const otherRequest = requests.find((request) => request.params.workspaceId === "other");
 		if (!otherRequest) throw new Error("missing independent layout request");
 		otherRequest.resolve({
-			snapshot: {
-				workspaceId: "other",
-				revision: 2,
-				document: otherRequest.params.document,
+			status: "accepted",
+			payload: {
+				snapshot: {
+					workspaceId: "other",
+					revision: 2,
+					document: otherRequest.params.document,
+				},
+				mutationId: otherRequest.params.mutationId,
 			},
-			mutationId: otherRequest.params.mutationId,
 		});
 		await independent;
 
@@ -161,6 +170,161 @@ describe("synchronized layout store", () => {
 		expect(requests).toHaveLength(2);
 		expect(useAppStore.getState().layoutDocumentsByWorkspace.ws).toEqual(document("accepted"));
 		expect(useAppStore.getState().layoutPendingByWorkspace.ws).toEqual([]);
+	});
+
+	test("two dependent commits succeed in captured revision order", async () => {
+		const requests: PendingRequest[] = [];
+		setLayoutReplaceRequesterForTests(
+			(params) =>
+				new Promise((resolve, reject) => {
+					requests.push({ params, resolve, reject });
+				}),
+		);
+		useAppStore.getState().installLayoutSnapshot(snapshot(1, "accepted"));
+		const first = commitWorkspaceLayout("ws", document("first"));
+		const second = commitWorkspaceLayout("ws", document("second"));
+		await Bun.sleep(0);
+		expect(requests).toHaveLength(1);
+		expect(requests[0]?.params.expectedRevision).toBe(1);
+		const firstRequest = requests[0];
+		if (!firstRequest) throw new Error("missing first request");
+		firstRequest.resolve({
+			status: "accepted",
+			payload: {
+				snapshot: { workspaceId: "ws", revision: 2, document: firstRequest.params.document },
+				mutationId: firstRequest.params.mutationId,
+			},
+		});
+		await first;
+		await Bun.sleep(0);
+
+		expect(requests).toHaveLength(2);
+		const secondRequest = requests[1];
+		if (!secondRequest) throw new Error("missing second request");
+		expect(secondRequest.params.expectedRevision).toBe(2);
+		secondRequest.resolve({
+			status: "accepted",
+			payload: {
+				snapshot: { workspaceId: "ws", revision: 3, document: secondRequest.params.document },
+				mutationId: secondRequest.params.mutationId,
+			},
+		});
+		await expect(second).resolves.toEqual(snapshot(3, "second"));
+		const state = useAppStore.getState();
+		expect(state.layoutSnapshotsByWorkspace.ws).toEqual(snapshot(3, "second"));
+		expect(state.layoutDocumentsByWorkspace.ws).toEqual(document("second"));
+		expect(state.layoutPendingByWorkspace.ws).toEqual([]);
+	});
+
+	test("a remote write between dependent commits conflicts and cancels later projections", async () => {
+		const requests: PendingRequest[] = [];
+		setLayoutReplaceRequesterForTests(
+			(params) =>
+				new Promise((resolve, reject) => {
+					requests.push({ params, resolve, reject });
+				}),
+		);
+		useAppStore.getState().installLayoutSnapshot(snapshot(1, "accepted"));
+		const first = commitWorkspaceLayout("ws", document("first"));
+		const conflicting = commitWorkspaceLayout("ws", document("conflicting")).then(
+			() => "fulfilled",
+			(error: Error) => error,
+		);
+		const dependent = commitWorkspaceLayout("ws", document("dependent")).then(
+			() => "fulfilled",
+			(error: Error) => error,
+		);
+		await Bun.sleep(0);
+		const firstRequest = requests[0];
+		if (!firstRequest) throw new Error("missing first request");
+		firstRequest.resolve({
+			status: "accepted",
+			payload: {
+				snapshot: { workspaceId: "ws", revision: 2, document: firstRequest.params.document },
+				mutationId: firstRequest.params.mutationId,
+			},
+		});
+		const remote = snapshot(3, "remote");
+		useAppStore.getState().applyLayoutChanged({ snapshot: remote, mutationId: "remote-client" });
+		await first;
+		await Bun.sleep(0);
+
+		expect(requests).toHaveLength(2);
+		const staleRequest = requests[1];
+		if (!staleRequest) throw new Error("missing stale dependent request");
+		expect(staleRequest.params.expectedRevision).toBe(2);
+		staleRequest.resolve({ status: "conflict", current: remote });
+
+		const conflictResult = await conflicting;
+		expect(conflictResult).toBeInstanceOf(Error);
+		expect((conflictResult as Error).name).toBe("LayoutCommitConflictError");
+		const dependentResult = await dependent;
+		expect(dependentResult).toBeInstanceOf(Error);
+		expect((dependentResult as Error).name).toBe("SupersededLayoutCommitError");
+		expect(requests).toHaveLength(2);
+		const state = useAppStore.getState();
+		expect(state.layoutSnapshotsByWorkspace.ws).toEqual(remote);
+		expect(state.layoutDocumentsByWorkspace.ws).toEqual(document("remote"));
+		expect(state.layoutPendingByWorkspace.ws).toEqual([]);
+		expect(state.toasts).toEqual([]);
+	});
+
+	test("a delayed conflict response cannot regress a newer accepted broadcast", () => {
+		const store = useAppStore.getState();
+		store.installLayoutSnapshot(snapshot(1, "accepted"));
+		store.beginLayoutCommit("ws", document("optimistic"), "mine");
+		const newer = snapshot(3, "newer-remote");
+		store.applyLayoutChanged({ snapshot: newer, mutationId: "remote-client" });
+
+		store.applyLayoutConflict("ws", "mine", snapshot(2, "conflict-time-current"));
+		const state = useAppStore.getState();
+		expect(state.layoutSnapshotsByWorkspace.ws).toEqual(newer);
+		expect(state.layoutDocumentsByWorkspace.ws).toEqual(newer.document);
+		expect(state.layoutPendingByWorkspace.ws).toEqual([]);
+	});
+
+	test("a delayed absent conflict cannot erase a creation broadcast that overtook it", () => {
+		const store = useAppStore.getState();
+		store.beginLayoutCommit("ws", document("optimistic-create"), "mine");
+		const created = snapshot(1, "remote-create");
+		store.applyLayoutChanged({ snapshot: created, mutationId: "remote-client" });
+
+		store.applyLayoutConflict("ws", "mine", null);
+		const state = useAppStore.getState();
+		expect(state.layoutSnapshotsByWorkspace.ws).toEqual(created);
+		expect(state.layoutDocumentsByWorkspace.ws).toEqual(created.document);
+		expect(state.layoutPendingByWorkspace.ws).toEqual([]);
+	});
+
+	test("a conflict can install authoritative absence and rolls back the optimistic document", async () => {
+		let pending:
+			| {
+					params: LayoutReplaceParams;
+					resolve: (result: LayoutReplaceResult) => void;
+			  }
+			| undefined;
+		setLayoutReplaceRequesterForTests(
+			(params) =>
+				new Promise((resolve) => {
+					pending = { params, resolve };
+				}),
+		);
+		useAppStore.getState().installLayoutSnapshot(snapshot(1, "stale-local"));
+		const committed = commitWorkspaceLayout("ws", document("optimistic")).then(
+			() => "fulfilled",
+			(error: Error) => error,
+		);
+		await Bun.sleep(0);
+		if (!pending) throw new Error("missing layout request");
+		pending.resolve({ status: "conflict", current: null });
+		const result = await committed;
+		expect(result).toBeInstanceOf(Error);
+		expect((result as Error).name).toBe("LayoutCommitConflictError");
+		const state = useAppStore.getState();
+		expect(state.layoutSnapshotsByWorkspace.ws).toBeUndefined();
+		expect(state.layoutDocumentsByWorkspace.ws).toBeUndefined();
+		expect(state.layoutPendingByWorkspace.ws).toEqual([]);
+		expect(state.toasts).toEqual([]);
 	});
 
 	test("a matching broadcast wins over a lost response without a false rollback", async () => {
@@ -180,6 +344,7 @@ describe("synchronized layout store", () => {
 		const committed = commitWorkspaceLayout("ws", document("next"));
 		await Bun.sleep(0);
 		if (!pending) throw new Error("missing layout request");
+		expect(pending.params.expectedRevision).toBe(1);
 		const accepted = {
 			workspaceId: "ws",
 			revision: 2,
@@ -206,6 +371,7 @@ describe("synchronized layout store", () => {
 	test("a failed first seed removes its unaccepted optimistic document", () => {
 		const store = useAppStore.getState();
 		store.beginLayoutCommit("ws", document("seed"), "seed");
+		expect(useAppStore.getState().layoutPendingByWorkspace.ws?.[0]?.expectedRevision).toBeNull();
 		store.rejectLayoutCommit("ws", "seed");
 		expect(useAppStore.getState().layoutDocumentsByWorkspace.ws).toBeUndefined();
 		expect(useAppStore.getState().layoutRemoteEpochByWorkspace.ws).toBe(1);
