+++ apps/web/src/shell/legacySelection/index.ts
@@ -1,0 +1,1 @@
+export * from "./legacySelection";
